+++ src/test/java/com/nexus/sion/feature/auth/command/application/service/AuthServiceImplTest.java
@@ -25,7 +25,7 @@ import com.nexus.sion.feature.auth.command.application.dto.response.TokenRespons
 import com.nexus.sion.feature.auth.command.domain.aggregate.RefreshToken;
 import com.nexus.sion.feature.member.command.domain.aggregate.entity.Member;
 import com.nexus.sion.feature.member.command.domain.aggregate.enums.MemberRole;
-import com.nexus.sion.feature.member.command.repository.MemberRepository;
+import com.nexus.sion.feature.member.command.domain.repository.MemberRepository;
 import com.nexus.sion.security.jwt.JwtTokenProvider;
 
 class AuthServiceImplTest {
