+++ src/test/java/com/nexus/sion/feature/member/command/MemberCommandIntegrationTest.java
@@ -32,7 +32,7 @@ import com.nexus.sion.feature.member.command.application.service.MemberCommandSe
 import com.nexus.sion.feature.member.command.domain.aggregate.entity.Member;
 import com.nexus.sion.feature.member.command.domain.aggregate.enums.MemberRole;
 import com.nexus.sion.feature.member.command.domain.aggregate.enums.MemberStatus;
-import com.nexus.sion.feature.member.command.repository.MemberRepository;
+import com.nexus.sion.feature.member.command.domain.repository.MemberRepository;
 
 @SpringBootTest
 @AutoConfigureMockMvc
